declaration.h: Convert enum BUILTIN to enum class

+++ src/dmd/func.d
@@ -64,7 +64,7 @@ enum ILS : int
     yes,                 /// can inline
 }
 
-enum BUILTIN : int
+enum BUILTIN : byte
 {
     unknown = -1,    /// not known if this is a builtin
     unimp,           /// this is not a builtin
